Added support for new WriteNode
This feature allows user to write content to a file. The files will be saved in a folder located as a sibling to the rivet project file with a name constructed by replacing "rivet-project" with "rivet-files". The path specified in the node will then be treated as a relative path.

The user can choose to either overwrite if exists, or to save with a unique name. Content will be saved as text or binary according to choices.

The path also supports resolving the extension by using '{{ext}}' which, for images or audio, will take the mediatype if available. Text will take 'txt' in all cases.